# src/web/openapi/running-performance.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RunningPerformance.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/exercises/{id}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "operationId": "GetExercise",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseMediaResponse": {
        "required": [
          "id",
          "position",
          "assetUri",
          "altText",
          "mimeType",
          "source",
          "author",
          "license",
          "sha256",
          "presentationSex",
          "widthPx",
          "heightPx"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "assetUri": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "author": {
            "type": [
              "null",
              "string"
            ]
          },
          "license": {
            "type": "string"
          },
          "sha256": {
            "type": [
              "null",
              "string"
            ]
          },
          "presentationSex": {
            "type": "string"
          },
          "widthPx": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "heightPx": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ExerciseResponse": {
        "required": [
          "id",
          "slug",
          "canonicalName",
          "movementPattern",
          "equipment",
          "status",
          "revision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "canonicalName": {
            "type": "string"
          },
          "movementPattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "equipment": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "revision": {
            "$ref": "#/components/schemas/ExerciseRevisionResponse"
          }
        }
      },
      "ExerciseRevisionResponse": {
        "required": [
          "id",
          "versionNumber",
          "displayName",
          "briefDescription",
          "setup",
          "execution",
          "safetyCues",
          "media"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "displayName": {
            "type": "string"
          },
          "briefDescription": {
            "type": "string"
          },
          "setup": {
            "type": "string"
          },
          "execution": {
            "type": "string"
          },
          "safetyCues": {
            "type": "string"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseMediaResponse"
            }
          }
        }
      },
...
    }
  },
...
}